### What does this do?

Since gocron requires go1.21.0+, we can replace "golang.org/x/exp/maps" and "golang.org/x/exp/slices" with the standard libraries "maps" and "slices". BTW, "golang.org/x/exp/slices" now is only a wrapper of "slice".


### Which issue(s) does this PR fix/relate to?


### List any changes that modify/break current functionality


### Have you included tests for your changes?


### Did you document any new/modified functionality?

### Notes

"removeSliceDuplicatesInt" could also get some performance benifits:

```go
func BenchmarkRemoveSliceDuplicatesInt(b *testing.B) {
	for range b.N {
		removeSliceDuplicatesInt([]int{1, 4, 5, 4, 6, 2, 3})
	}
}
```

```text
goos: windows
goarch: amd64
pkg: bench
cpu: Intel(R) Core(TM) i5-10200H CPU @ 2.40GHz
                           │     old      │                 new                 │
                           │    sec/op    │   sec/op     vs base                │
RemoveSliceDuplicatesInt-8   285.00n ± 1%   29.75n ± 0%  -89.56% (p=0.000 n=10)

                           │    old     │                new                 │
                           │    B/op    │   B/op     vs base                 │
RemoveSliceDuplicatesInt-8   96.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=10)

                           │    old     │                 new                 │
                           │ allocs/op  │ allocs/op   vs base                 │
RemoveSliceDuplicatesInt-8   3.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
```

As an added benefit, there is now no need to sort the results of “removeSliceDuplicatesInt” because they are already sorted.